web-Proyecto/VetifyBooking/requirements.txt
Django>=6.0,<7.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
xhtml2pdf
